+++ NEWS.md
@@ -7,6 +7,9 @@
   anyway, because PPM is in a consistent state w.r.t. metadata and
   package files (https://github.com/r-lib/pak/issues/623).
 
+* `parse_packages()` now does not throw a warning for empty `PACKAGES*`
+  files (#107).
+
 # pkgcache 2.2.2
 
 * pkgcache now treats R 4.5.x (current R-devel) macOS binaries